+++ NEWS.md
@@ -2,8 +2,11 @@
 
 ## New functions
 
-- Added `stdout` utility, which helps parse output from `shell` return.
 - Migrated `gitCurrentBranch` and `gitDefaultBranch` from r-koopa here.
+
+## Minor changes
+
+- `shell`: Added `returnStdout` argument, which helps parse stdout.
 
 # AcidBase 0.6.0 (2022-04-29)